base connection for tests
